removed some static_casts

+++ src/kanji/RadicalData.cpp
@@ -12,7 +12,7 @@ void RadicalData::load(const std::filesystem::path& file) {
     longNameCol("LongName"), readingCol("Reading");
   for (ColumnFile f(file, {numberCol, nameCol, longNameCol, readingCol});
        f.nextRow();) {
-    const auto radicalNumber = static_cast<size_t>(f.getSize(numberCol));
+    const auto radicalNumber = f.getSize(numberCol);
     if (radicalNumber != f.currentRow())
       f.error("radicals must be ordered by 'number'");
     std::stringstream radicals(f.get(nameCol));
